wip: fixing pallet_council tests
