+++ bin/node/pallets/pallet-council/Cargo.toml
@@ -21,6 +21,7 @@ frame-benchmarking = { version = "4.0.0-dev", default-features = false, path = "
 frame-support = { version = "4.0.0-dev", default-features = false, path = "../../../../frame/support" }
 frame-system = { version = "4.0.0-dev", default-features = false, path = "../../../../frame/system" }
 pallet-assets = { version = "4.0.0-dev", default-features = false, path = "../../../../frame/assets" }
+pallet-balances = { version = "4.0.0-dev", default-features = false, path = "../../../../frame/balances" }
 
 [features]
 default = ["std"]
